Convert `substitute` to TypeScript.
